Move interfaces to their own file and inforce type imports
