+++ .eslintrc
@@ -72,7 +72,8 @@
                 "@typescript-eslint/prefer-string-starts-ends-with": "error",
                 "@typescript-eslint/switch-exhaustiveness-check": "error",
                 // We intentionally use this so that constructs can declare their command handlers
-                "@typescript-eslint/unbound-method": "off"
+                "@typescript-eslint/unbound-method": "off",
+                "@typescript-eslint/consistent-type-imports": "error"
             }
         }
     ]
