+++ src/constructs/abstracts/AwsConstruct.ts
@@ -1,6 +1,6 @@
 import { Construct as CdkConstruct } from "@aws-cdk/core";
-import { AwsProvider } from "@lift/providers";
-import { ConstructInterface } from "@lift/constructs";
+import type { AwsProvider } from "@lift/providers";
+import type { ConstructInterface } from "@lift/constructs";
 
 export abstract class AwsConstruct extends CdkConstruct implements ConstructInterface {
     static create<C extends AwsConstruct = AwsConstruct>(
