+++ src/constructs/aws/DatabaseDynamoDBSingleTable.ts
@@ -1,7 +1,8 @@
-import { Construct as CdkConstruct, CfnOutput, Fn, Stack } from "@aws-cdk/core";
+import type { Construct as CdkConstruct } from "@aws-cdk/core";
+import { CfnOutput, Fn, Stack } from "@aws-cdk/core";
 import { AttributeType, BillingMode, StreamViewType, Table } from "@aws-cdk/aws-dynamodb";
-import { FromSchema } from "json-schema-to-ts";
-import { AwsProvider } from "@lift/providers";
+import type { FromSchema } from "json-schema-to-ts";
+import type { AwsProvider } from "@lift/providers";
 import { AwsConstruct } from "@lift/constructs/abstracts";
 import { PolicyStatement } from "../../CloudFormation";
