+++ src/constructs/aws/Webhook.ts
@@ -1,10 +1,11 @@
-import { Construct as CdkConstruct, CfnOutput, Fn } from "@aws-cdk/core";
+import type { Construct as CdkConstruct } from "@aws-cdk/core";
+import { CfnOutput, Fn } from "@aws-cdk/core";
 import { CfnAuthorizer, CfnIntegration, CfnRoute, HttpApi } from "@aws-cdk/aws-apigatewayv2";
 import { Function } from "@aws-cdk/aws-lambda";
 import { EventBus } from "@aws-cdk/aws-events";
-import { FromSchema } from "json-schema-to-ts";
+import type { FromSchema } from "json-schema-to-ts";
 import { PolicyDocument, PolicyStatement, Role, ServicePrincipal } from "@aws-cdk/aws-iam";
-import { AwsProvider } from "@lift/providers";
+import type { AwsProvider } from "@lift/providers";
 import { AwsConstruct } from "@lift/constructs/abstracts";
 import ServerlessError from "../../utils/error";
