+++ src/constructs/aws/queue/sqs.ts
@@ -1,4 +1,4 @@
-import {
+import type {
     DeleteMessageBatchRequest,
     DeleteMessageBatchResult,
     Message,
@@ -7,7 +7,7 @@ import {
     SendMessageBatchRequest,
     SendMessageBatchResult,
 } from "aws-sdk/clients/sqs";
-import { AwsProvider } from "@lift/providers";
+import type { AwsProvider } from "@lift/providers";
 import { log } from "../../../utils/logger";
 import { sleep } from "../../../utils/sleep";
