+++ src/plugin.ts
@@ -1,13 +1,14 @@
 import { flatten, get, has, merge } from "lodash";
 import chalk from "chalk";
-import { AwsIamPolicyStatements } from "@serverless/typescript";
+import type { AwsIamPolicyStatements } from "@serverless/typescript";
 import * as path from "path";
 import { readFileSync } from "fs";
 import { dump } from "js-yaml";
 import { DefaultTokenResolver, Lazy, StringConcat, Tokenization } from "@aws-cdk/core";
-import { FromSchema } from "json-schema-to-ts";
-import { AwsProvider, ProviderInterface, StaticProviderInterface, StripeProvider } from "@lift/providers";
-import { ConstructInterface, StaticConstructInterface } from "@lift/constructs";
+import type { FromSchema } from "json-schema-to-ts";
+import type { ProviderInterface, StaticProviderInterface } from "@lift/providers";
+import { AwsProvider, StripeProvider } from "@lift/providers";
+import type { ConstructInterface, StaticConstructInterface } from "@lift/constructs";
 import type {
     CommandsDefinition,
     DeprecatedVariableResolver,
