+++ src/providers/ProviderInterface.ts
@@ -1,0 +1,5 @@
+import type { ConstructInterface } from "@lift/constructs";
+
+export interface ProviderInterface {
+    create(type: string, id: string): ConstructInterface;
+}
