+++ src/providers/StaticProviderInterface.ts
@@ -1,0 +1,6 @@
+import type { StaticConstructInterface } from "@lift/constructs";
+
+export interface StaticProviderInterface {
+    getConstructClass(type: string): StaticConstructInterface | undefined;
+    getAllConstructClasses(): StaticConstructInterface[];
+}
