+++ src/utils/s3-sync.ts
@@ -1,4 +1,4 @@
-import {
+import type {
     DeleteObjectsOutput,
     DeleteObjectsRequest,
     ListObjectsV2Output,
@@ -14,7 +14,7 @@ import * as crypto from "crypto";
 import { lookup } from "mime-types";
 import { chunk, flatten } from "lodash";
 import chalk from "chalk";
-import { AwsProvider } from "@lift/providers";
+import type { AwsProvider } from "@lift/providers";
 
 const readdir = util.promisify(fs.readdir);
 const stat = util.promisify(fs.stat);
